Move unused classes back to inner
